requirements.txt
streamlit>=1.28
pandas
plotly
openpyxl
requests
python-dotenv
openai
openai>=1.0.0.
qrcode[pil]
Pillow
nump

